[feat] Improve UI layout and assets (#64)
* [feat] Improve UI layout and assets

Improve the UI layout and assets to better align with the Greek Mythology theme and reduce the amount of wasted space.

Previous implementations had the players and active God consuming 512px x 128px each, resulting in significant screen real estate being used in 4 player games.

New implementation moves all of the information to a banner across the top of the screen. This ensures consistent presentation, regardless of the number of players, and reduces overall complexity of the layout significantly.